Update pomodoro.css
I added those two classes (the one for the start and stop buttons and the one for the work and rest buttons) and removed the styles for the id's of these buttons. It is simpler to make changes to the styles of the buttons this way, since those four buttons always change in the pairs defined by these two new classes.